Replace default exports with named exports
