+++ .changeset/hot-clouds-applaud.md
@@ -1,0 +1,5 @@
+---
+"saleor-app-search": minor
+---
+
+Removed search index preview page. It can be easily accessed at Algolia itself.
