+++ .changeset/new-mice-remember.md
@@ -1,0 +1,5 @@
+---
+"saleor-app-search": patch
+---
+
+Update Next and Sentry
