+++ .changeset/thick-dots-live.md
@@ -1,0 +1,13 @@
+---
+"eslint-config-saleor": patch
+"saleor-app-emails-and-messages": patch
+"saleor-app-data-importer": patch
+"saleor-app-products-feed": patch
+"saleor-app-monitoring": patch
+"saleor-app-klaviyo": patch
+"saleor-app-search": patch
+"saleor-app-slack": patch
+"saleor-app-taxes": patch
+---
+
+Replace "export default" with named exports
