+++ apps/data-importer/src/pages/importer.tsx
@@ -3,7 +3,7 @@ import React, { ComponentProps } from "react";
 import { Container, Divider } from "@material-ui/core";
 import { Button, makeStyles, PageTab, PageTabs, SaleorTheme } from "@saleor/macaw-ui";
 import { CustomersImporterView } from "../modules/customers/customers-importer-nuvo/customers-importer-view";
-import GraphQLProvider from "../providers/GraphQLProvider";
+import { GraphQLProvider } from "../providers/GraphQLProvider";
 import { actions, useAppBridge } from "@saleor/app-sdk/app-bridge";
 
 type Tab = "customers";
