+++ apps/klaviyo/src/pages/api/webhooks/customer-created.ts
@@ -2,45 +2,69 @@ import { NextWebhookApiHandler, SaleorAsyncWebhook } from "@saleor/app-sdk/handl
 import { gql } from "urql";
 
 import {
-  OrderCreatedWebhookPayloadFragment,
-  UntypedOrderCreatedDocument,
-} from "../../../generated/graphql";
+  CustomerCreatedWebhookPayloadFragment,
+  UntypedCustomerCreatedDocument,
+} from "../../../../generated/graphql";
 import { createClient } from "../../../lib/graphql";
-import Klaviyo from "../../../lib/klaviyo";
+import { Klaviyo } from "../../../lib/klaviyo";
 import { createSettingsManager } from "../../../lib/metadata";
-import { saleorApp } from "../../../saleor-app";
+import { saleorApp } from "../../../../saleor-app";
 
-const OrderCreatedWebhookPayload = gql`
-  fragment OrderCreatedWebhookPayload on OrderCreated {
-    order {
-      ...OrderFragment
+const CustomerCreatedWebhookPayload = gql`
+  fragment CustomerCreatedWebhookPayload on CustomerCreated {
+    user {
+      __typename
+      id
+      defaultShippingAddress {
+        ...AddressFragment
+      }
+      defaultBillingAddress {
+        ...AddressFragment
+      }
+      addresses {
+        ...AddressFragment
+      }
+      privateMetadata {
+        ...MetadataFragment
+      }
+      metadata {
+        ...MetadataFragment
+      }
+      email
+      firstName
+      lastName
+      isActive
+      dateJoined
+      languageCode
     }
   }
 `;
 
-export const OrderCreatedGraphqlSubscription = gql`
-  ${OrderCreatedWebhookPayload}
-  subscription OrderCreated {
+export const CustomerCreatedGraphqlSubscription = gql`
+  ${CustomerCreatedWebhookPayload}
+  subscription CustomerCreated {
     event {
-      ...OrderCreatedWebhookPayload
+      ...CustomerCreatedWebhookPayload
     }
   }
 `;
 
-export const orderCreatedWebhook = new SaleorAsyncWebhook<OrderCreatedWebhookPayloadFragment>({
-  name: "Order Created",
-  webhookPath: "api/webhooks/order-created",
-  event: "ORDER_CREATED",
-  apl: saleorApp.apl,
-  query: UntypedOrderCreatedDocument,
-});
+export const customerCreatedWebhook = new SaleorAsyncWebhook<CustomerCreatedWebhookPayloadFragment>(
+  {
+    name: "Customer Created",
+    webhookPath: "api/webhooks/customer-created",
+    event: "CUSTOMER_CREATED",
+    apl: saleorApp.apl,
+    query: UntypedCustomerCreatedDocument,
+  }
+);
 
-const handler: NextWebhookApiHandler<OrderCreatedWebhookPayloadFragment> = async (
+const handler: NextWebhookApiHandler<CustomerCreatedWebhookPayloadFragment> = async (
   req,
   res,
   context
 ) => {
-  console.debug("orderCreatedWebhook handler called");
+  console.debug("customerCreatedWebhook handler called");
 
   const { payload, authData } = context;
   const { saleorApiUrl, token, appId } = authData;
@@ -48,14 +72,14 @@ const handler: NextWebhookApiHandler<OrderCreatedWebhookPayloadFragment> = async
   const settings = createSettingsManager(client, appId);
 
   const klaviyoToken = await settings.get("PUBLIC_TOKEN");
-  const klaviyoMetric = await settings.get("ORDER_CREATED_METRIC");
+  const klaviyoMetric = await settings.get("CUSTOMER_CREATED_METRIC");
 
   if (!klaviyoToken || !klaviyoMetric) {
     console.debug("Request rejected - app not configured");
     return res.status(400).json({ success: false, message: "App not configured." });
   }
 
-  const { userEmail } = payload.order || {};
+  const userEmail = payload.user?.email;
 
   if (!userEmail) {
     console.debug("Request rejected - missing user email");
@@ -68,6 +92,7 @@ const handler: NextWebhookApiHandler<OrderCreatedWebhookPayloadFragment> = async
   if (klaviyoResponse.status !== 200) {
     const klaviyoMessage = ` Message: ${(await klaviyoResponse.json())?.message}.` || "";
     console.debug("Klaviyo returned error: ", klaviyoMessage);
+
     return res.status(500).json({
       success: false,
       message: `Klaviyo API responded with status ${klaviyoResponse.status}.${klaviyoMessage}`,
@@ -78,7 +103,7 @@ const handler: NextWebhookApiHandler<OrderCreatedWebhookPayloadFragment> = async
   return res.status(200).json({ success: true, message: "Message sent!" });
 };
 
-export default orderCreatedWebhook.createHandler(handler);
+export default customerCreatedWebhook.createHandler(handler);
 
 export const config = {
   api: {
