+++ apps/klaviyo/src/pages/api/webhooks/order-fully-paid.ts
@@ -4,11 +4,11 @@ import { gql } from "urql";
 import {
   OrderFullyPaidWebhookPayloadFragment,
   UntypedOrderFullyPaidDocument,
-} from "../../../generated/graphql";
+} from "../../../../generated/graphql";
 import { createClient } from "../../../lib/graphql";
-import Klaviyo from "../../../lib/klaviyo";
+import { Klaviyo } from "../../../lib/klaviyo";
 import { createSettingsManager } from "../../../lib/metadata";
-import { saleorApp } from "../../../saleor-app";
+import { saleorApp } from "../../../../saleor-app";
 
 const OrderFullyPaidWebhookPayload = gql`
   fragment OrderFullyPaidWebhookPayload on OrderFullyPaid {
