+++ apps/search/package.json
@@ -20,7 +20,7 @@
     "@saleor/app-sdk": "0.37.1",
     "@saleor/apps-shared": "workspace:*",
     "@saleor/macaw-ui": "0.7.2",
-    "@sentry/nextjs": "^7.31.1",
+    "@sentry/nextjs": "^7.46.0",
     "@types/debug": "^4.1.7",
     "@urql/exchange-auth": "^1.0.0",
     "algoliasearch": "4.14.2",
@@ -29,7 +29,7 @@
     "graphql": "^16.6.0",
     "graphql-tag": "^2.12.6",
     "instantsearch.css": "^7.4.5",
-    "next": "13.1.6",
+    "next": "13.2.4",
     "next-urql": "4.0.0",
     "react": "18.2.0",
     "react-dom": "18.2.0",
