+++ apps/search/src/providers/GraphQLProvider.tsx
@@ -4,7 +4,7 @@ import { Provider } from "urql";
 
 import { createClient } from "../lib/graphql";
 
-function GraphQLProvider(props: PropsWithChildren<{}>) {
+export function GraphQLProvider(props: PropsWithChildren<{}>) {
   const { appBridgeState } = useAppBridge();
   const saleorApiUrl = appBridgeState?.saleorApiUrl!;
 
@@ -18,5 +18,3 @@ function GraphQLProvider(props: PropsWithChildren<{}>) {
 
   return <Provider value={client} {...props} />;
 }
-
-export default GraphQLProvider;
