+++ apps/slack/src/hooks/useAppApi.ts
@@ -11,7 +11,7 @@ interface UseFetchProps {
 }
 
 // This hook is meant to be used mainly for internal API calls
-const useAppApi = <D>({ url, options, skip }: UseFetchProps) => {
+export const useAppApi = <D>({ url, options, skip }: UseFetchProps) => {
   const { appBridgeState } = useAppBridge();
 
   const [data, setData] = useState<D>();
@@ -60,5 +60,3 @@ const useAppApi = <D>({ url, options, skip }: UseFetchProps) => {
 
   return { data, error, loading };
 };
-
-export default useAppApi;
