+++ apps/taxes/src/pages/_app.tsx
@@ -12,7 +12,7 @@ import {
 import React, { PropsWithChildren, useEffect } from "react";
 import { AppProps } from "next/app";
 
-import GraphQLProvider from "../providers/GraphQLProvider";
+import { GraphQLProvider } from "../providers/GraphQLProvider";
 import { ThemeSynchronizer } from "../lib/theme-synchronizer";
 import { trpcClient } from "../modules/trpc/trpc-client";
 import { NoSSRWrapper } from "../lib/no-ssr-wrapper";
